src-be/src/main/java/com/example/service/impl/UserServiceImpl.java
package com.example.service.impl;

import com.example.dto.UserDto;
import com.example.entity.User;
import com.example.mapper.UserMapper;
import com.example.repository.UserRepository;
import com.example.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import javax.persistence.EntityNotFoundException;

@Service
@RequiredArgsConstructor
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;
    private final UserMapper userMapper;

    @Override
    public User getByEmail(String email) {
        return userRepository.findByEmail(email)
                .orElseThrow(() -> new EntityNotFoundException(String.format("Can't find user with emil: %s", email)));
    }

    @Override
    public UserDto getUserByEmailAsDto(String email) {
        return userMapper.userToUserDto(getByEmail(email));
    }

}
